Fixed issue 30134: Error executing a process definition if a param. is numeric
The problem was that if a parameter was defined as default with a numeric value,
an error was got.

The problem was in the "ob-parameter-window-view.js", there is a function called
getTypeSafeValue which was not taking into account the fact of having the "value"
argument as a numeric value. That was why it was not working fine.

To fix this a new condition has been added. Now it is checked if the "value" parameter
is a number.

--HG--
extra : rebase_source : bfbebcddfab2f04bcc1ce7d762fa5b2780f97777