Fixed issue 30249: Error get when the Set New Currency process def. is executed
The problem was that a error was got when clicking the "Done" button of the "Set  New Currency"
process definition.

"Uncaught TypeError: Cannot read property '_handleSelect' of null"

The problem was in the "ob-parameter-window-view.js" in the handleResponse function.
It was executing "this.showProcessing(false)" even when the "retryExecution" parameter
was false.
Now the "this.showProcessing(false)" is executed only if the "retryExecution" parameter is true.

This affected to the process definition execution because as we were executing the "this.showProcessing(false)"
even when the retryExecution parameter was false, an ob-selector was deleted and after that, it was trying to
get the "._handleSelect" of that selector.