Related to issue 30057: code review preview
Removed flushes from RoleEventHandler because they create conflicts. As a consequence we lose the ability to clear the session, but this shouldn't be a problem because it's very unlikely to have environments with great amount of organizations.
In RoleEventHandler we only create access to * organization when access level is Client (before this changeset it created also records for other organizations and it is useless).
Reverted changes for Initial Client/Organization Setup. RoleEventHandler won't do anything if executed from an Initial Client/Organization Setup (this change is a way to isolate the problems reported at #30253)
Added log4j to RoleEventHandler.
Removed admin mode from RoleEventHandler because it's not needed (the user always has access t the entities involved in the process).

--HG--
extra : transplant_source : 2%E2%F1%F3%2A%AB%C3ye%BDE%FB%E2H%A4.h%2B%9E%E1