Fixed issue 30164: property order and property line are not retrieved quotation
The problem was that the property order and property line are not retrieved when
quotation is loaded and the description button doesn't appear when the
quotations was reactivated.

The solution is to add these parameters in the hql retrieve query and set in the
order.js model. The description button must be shown when the quotation is
reactivated.

Now, yo can change the description of the line and the receipt properties of a
reactivated quotation.