Fixed issue 30269: Focus keeper is enabled when edit tab is opened.
When Terminal Types are marked to work with external keyboards, the focus keeper is the engine which captures the inputs from this external devices. To allow web pos to catch these events the focus keeper must be enabled.

Before this fix, focus keeper was disabled when this tab is opened.
After this fix, the focus keeper is enabled when edit tab is opened, doing it if the user comes from a situation where the focus keeper is disabled it will be enabled and it will work fine. If the user comes from a situation where the focus keeper is already enabled it will continue working.