Backout commit 4799 / issue 29463
The proposed solution doesn't fix the real problem of the issue 29463.

The problem is not that you can done a payment with no cash and the return
of change by card method, the real problem is that the data generated in this
case is not saved well in the backend.