Fixes issue 30274: uuid appear in saved views with multiple criteria on a field
When a view is saved while any of its filters are currently using the 'id' filter type, the id and identifier of the selected options are stored in a cache. If we have multiple criteria on one field, an internal loop is executed to retrieve all the criterias in the getFKFilterAuxiliaryCache function.

The problem was that this internal loop was using the same variable as the external one to insert those criterias into the cache. This caused that the subsequent filters for next fields were not being handled properly as the condition of the external loop was not being fulfilled anymore due to the increment done in the internal loop.

This has been solved just using a new variable 'j' in the internal loop.