Fixed issue 30227: Cannot reactivate an invoice if it is paid with credit
When a receipt is created in web pos, and it is paid using credit, when cashup
is done, the invoice is created, but the invoice can not be reactivated.

The problem was that the payment was set as not Prepayment and, when in the back
office it's necessary to reactivate the invoice, the function
aprm_gen_paymentschedule_inv raise and error.

The solution is to set to true isPrepayment column if it's used credit.

Now, it's possible to Reactivate the invoice if the customer used credit