Fixes issue 30232: Costing record not referencing transaction
getRelatedTransactions method in AverageCostAdjustment.java was creating an HQL query with a wrong parenthesis. Because of that Order By clause did not appear in the query, what was causing wrong Costing entries.
Also, when this error happens, current costing will be updated with the new cost, instead of creating a new costing with a null transaction.