Fixed issue 30320: Touchscreens does not disable new customer button
Is is currently using this.setDisabled(), with no parameter. That set disabled property as undefined and once it is clicked on a touchscreen it executes tap button normally.
It is needed to send the boolean parameter to set the status of disabled. In this case this.setDisabled(true)

--HG--
extra : rebase_source : 7e658954b65849446ae70c1431f8f3278de891e5