Fixed issue 30267. Receipts search in WEB POS does not work with the char -
The problem was that the search process removes the char '-' and the filter from
the sql doesn't work as expected. This is a security measure to avoid sql
injection.

The solution is to add named parameters and change the filter value in the sql
string.

Now the filter works fine.