Fixes Issue 29314:Clicking select all and after that unselect all records in Credit to Use grid in AddPayment window has performance issues
The problem was that on every selection/deselection of record in creditUseGrid
it was triggering the execution of these three methods:
* OB.APRM.AddPayment.updateCreditTotal(view.theForm);
* OB.APRM.AddPayment.updateActualExpected(view.theForm);
* OB.APRM.AddPayment.distributeAmount(view, view.theForm, true);

Actually it is not necessary to execute these functions on every record, because
in the end they are updating total values, so just executing them after the
iteration of the last record selected/deselected it would be enough.

To implement this, 'userSelectAllRecords' and 'deselectAllRecords' ListGrid
methods have been overriden, to set a flag to true, in order to know when a
record is changed while clicking selected/deselect all records, and in this
case the three methods will be executed in the last record of the grid.
The 'OB.APRM.AddPayment.userSelectAllRecords' is executed only when clicking
in select all records checkbox, but 'OB.APRM.AddPayment.deselectAllRecords' is
executed while unchecking the previous checkbox and when fetching data in the
grid using 'invalidateCache()' method. In this last case these functions will
not be executed on every deselection neither in the last record, because they
are already invoked directly in 'invalidateCache()', so in this case they will
be avoided.

While executing invalidateCache() function to reload CreditUse grid it
internally calls to 'OB.APRM.AddPayment.deselectAllRecords' overriden method
that sets obaprmAllRecordsSelectedByUser property to true. This property is
only needed when manually the select/unselect all checkbox is clicked.In order
to avoid this in creditOnLoadGrid function, if the property
'obaprmAllRecordsSelectedByUser' exists, it will be deleted.

--HG--
extra : rebase_source : 501c3e83501950d55eedf167af525ca4cd3796b0