[pool] Fixes problems when re-start tomcat with new external connection pool
The problem was found when a configuration change tomcat to make some tests. In config/server.xml
disabled the following line: <Listener className="org.apache.catalina.core.JreMemoryLeakPreventionListener" />.
The problem that appeared was that somehow the pool was not being closed properly. This caused problems when
shutting down and restarting tomcat because "openbravo" has tried to get connections of the not properly closed pool.

It has been implemented and has been forced to make a close of the pool when the shutdown process of tomcat
starts. When tomcat shuts down, the pool has been closed properly even all active connections. Now, It has
fixed the problem.