fixes issue 30459: can execute modulescripts twice or more on install.source
It was not possible to configure modulescripts to be executed more than once during install.source of different versions. Now we send an empty hash map during install.source to the ModuleScriptHandler. This way we identify that we are executing the install.source task. Thus, in this case we will execute the modulescript depending on the value returned by its executeOnInstall() method

--HG--
extra : rebase_source : 9317b0447e71a195a9dd4c611abb0b116c793cc6