fixes issue 30204: prevent unpaged datasource request from manual sources
Now the unpaged datasource requests are controlled by the OBJSON_AllowUnpagedDatasourceManualRequest preference. If it is set to N (default) is not possible to perform these kind of requests. They can be enabled by setting this preference to Y.
In order to have a better performance on the datasource when requesting for the preference value, the value of the preference is being cached by the UnpagedRequestCachedPreference class