Fixes Issue 30490: Return from Customer Pick&Edit Lines allows to enter more quantity to return than the pending to be returned
Return document has negative movementqty so ReturnQtyOtherRM was negative.
It should be postive so unlike ReturnQty it is multiplied with -1 as in prev
Datasource prior to transforming it to HQL PE. Where clause which checks
MovementQty of shipment line with the ReturnQtyOtherRM was missing. Sum of
Shipmentline MovementQty and ReturnQtyOtherRM should not be zero. It is added
to fix the problem.