Fixes issue 30464: Openbravo ERP 3 doesn't work on Microsoft Edge
The application was crashing silently during the login process, because Smartclient sources were not identyfing properly the browser. This caused an error in the SC getStackTrace() method. In addition, the OB function getBrowserInfo() has been updated to identify the Microsft Edge browser in order to declare it as not supported for the time being

--HG--
extra : rebase_source : af3d6b0a21938524d0f5ffc4ce9c21d96d62c169