Related to issue 30470: code review improvements
The error message when manually posting the record was overwritten by the "@NotCalculatedCost@" string because the catch code always overwrites the message set in the correspondent Doc* class.
This was wrong, as the Doc* class should be in charge of setting the error message.
Now the message is only overwritten if it hasn't previously set by the correspondent Doc* class

--HG--
extra : transplant_source : %25%FEw%A2%87.P%03%7B%0D%EB%9D%82%0D%F38P%B8L%FF