Fixes issue 30561: UPPER is used inCustomQuerySelectorDatasource to build query
When the CustomQuerySelectorDatasource built the query to achieve case insensitive filtering, it used to use the LOWER function. We decided to change this, as UPPER is used much more often throughout the codebase. The fix has been checked with custom query selectors using both the iContains and the iStartsWith operators