Fixes issue 30571: Rows with mandatory image columns can be removed
When a row contains columns that use the ad_image reference, those images are removed when the row is removed in the RemoveImagesEventHandler. This is an event handler that is executed before the row is removed, so removing the image results in a query like this:

update my_table set my_image_field = null

This is a problem is the image column is mandatory.

There were two ways to fix this:
- Give support to event handlers that are executed after the row is removed.
- Set a dummy image to the deleted row, to avoid breaking the not null constraint.

The first option is very complex, so the first one has been chosen.

--HG--
extra : rebase_source : bb2f88da548e3b189f97c1244d36308f730f9664