Fixes issue 30607: The sales section of the cashup report is recovered from the already existing cashup information
The problem was that the cashup report was being generated on demand and we do not keep track of the cashupId to which a sale belonged before it is assigned to a new cashupId when it is paid.

The solution has been to retrieve the information from the already existing data.
This solucion could be extended to the other sections of the cashup report. We like the fact that is being regenerated because it has been a good source of catching issues

--HG--
extra : rebase_source : 329088de2ba791a5a2a5bd003ec50d30e260f62f