fixes issue 27247: Tomcat 8 is not supported
The changes are mainly related to the change in the behavior of the getRealPath method of the StandardContext class. Since Tomcat 8, the provided path should start with a slash. Also this method is now making use of a refactor done in the way the web resources are handled (https://tomcat.apache.org/tomcat-8.0-doc/config/resources.html).

--HG--
extra : rebase_source : 3059e3fe0cba47acbaa1a30415c087d8eb6d44b4