Fixes issue 30425: Partial payments in layaway through "Pay Open Tickets" option
The problem was that the partial payment of a layaway always creates a new
order, not taken into account that there is already a order created with the
same id and attributes.

The solution is to modify the newLayaway parameter to represent what it means in
the flow.

--HG--
extra : rebase_source : 00a3fa1a8b73b95c890c0295518dc1508599496e