related to issue 30761: Product characteristics popup does not work properly
If any FK filter is used, the values shown in the other filters are restricted by them. This was not
happening with the product characteristics popu. In this case the whole tree was shown.

To fix this some steps have been done.
First of all it was necessary to change the where clause to take into account the filters applied in the grid.
On the other hand it has been necessary to take into account when a filter was applied in the "Product Characteristics"
column, because if at first you choose a value of a product characteristic, the second time you open the popup of
product characteristics, it is not possible to see any other product characteristic but the one previously chosen. That
was not working fine, it only has to take into account the filters applied in the grid.
To fix this, the product characteristics criteria has been removed if it existed one, and these all changes make possible to
have the values of the product characteristics filtered applying the filters of the grid but not taking into account the value
chosen in the first moment in the product characteristics popup.