Fixes issue 30425: Partial payments in layaway through "Pay Open Tickets" option
The problem was that the partial payment of a layaway always creates a new
order, not taken into account that there is already a order created with the
same id and attributes.

The solution is to create a new check to see if the layaway is partially paid.

--HG--
extra : rebase_source : 310e0c958134559e0521441d6ea677ccdeb3fafe