Fixed bug 30735: code review improvements
Rewritten validation rule to be generic, so we can use it for any record (not only Price List Versions).
In case the role's user level is Client or Client+Organization, the validation also displays PLV for * organization even in the case the * organization is not in the list of the Role's Organizations. This is the way it works for normal WAD windows (like Price List).
Removed code related to user level = System, as this is not used by the finance flows

--HG--
extra : transplant_source : %DF%17%2C%11i%5E%28%F2k%B7T%83%A6x%FA%C7_%27%5B%BA