Fixes issue 30033: Can't create stock for product with attribute without stock
We will not be able to void a receipt with a product with a mandatory attribute set value and without stock, if client does not allow negative stock. We will also not be able to create an inventory to increase the stock of this product because we can not select this attribute set value with a selector, and Create Inventory Count List process does not insert lines with products without stock if product has a mandatory attribute.
Now, if we are voiding a Goods Receipt with any line with mandatory attribute set value and without stock, a new popup will allow us to automatically create a Physical Inventory to generate stock for this product and then void the receipt.

--HG--
extra : rebase_source : f723d9323c7f5c3854cb12ccd63006d13b426001