Fixed bug 30866: Match Statement Transactions prior to Bank Statement Line date
When searching for transactions, the HQL query transformer added a where clause to get only the transactions prior or equal to the bank statement line's transaction date.
There are cases where you want to match to transactions registered after the bank statement line's transaction date.
Before the refactor, there was a flag to hide/show transactions in the future. With the refactor that functionality was lost.

To implement it again, we have moved that restriction from the hql query transformer to a new default filter for this tab. Thus the user can remove the filter to show all the transactions regardless the transaction date.