Fixed issue 30778: Cash UP Partial crashes
The problem was that the cashup partial tries to start from the step 5
(step CashKeep) instead of the step 6 (step PostPrint&Close).

Some functionality has changed and the cashup partial doesn't show the real
info in the report and in the ticket.

--HG--
extra : rebase_source : 74248724c55d60521da972fe50e426faed745630