Fixes bug 30507:Clicking on a field on grid edit mode the value is not selected
The problem was when editing a row on grid view if you click on a field the value
is not selected. It is a regression in Smartclient library.

To resolve this problem it has been included the code of Smartclient that fixes
Smartclient's regression. Besides another minor problem it has been detected in
Smartclient library. It has been resolved this problem too.