Fixed issue 30812:Allow multiple filtering using tree reference.
It was not allowed the multiple filtering using a tree reference in a
pick and execute window's grid.

The ob-formitem-tree-filter.js has been changed for being able to do this.
The problem was in the "filterDialogCallback" funtion in the "this.grid.parentElement.setFilterEditorCriteria(criteria)"
It was not working fine in every case.
It was neccesary to change the criteria for having the following: It was neccesary to delete
the old criteria of the tree filter, taking into account the new criteria of the tree filter
and taking into account the criteria of the rest of the grid of the pick and Execute window.

The new function, "setCurrentCriterion" is the one that updates the criteria of the tree filter, it deletes the
old criteria of the filter and gets the new criteria of the tree filter.

--HG--
extra : rebase_source : b0a77e20e08c097c089d0fab4ce73295452b2054