Related to issue 30812: Improve the functions.
The function filterDialogCallback of the "ob-formitem-tree-filter.js" is improved.
This function calls the "updateCriterion" function which updates the criterion of
the tree filter deleting the old criterion and adding the new criterion.
The crieteria of the rest of the columns in the grid is taken into account.

--HG--
extra : transplant_source : %0F%F6%25%24Z%B9%9Eg%DF%D7X%AE%E8%CB%81%E9%E2%BD%A8%7F