Fixed bug 30926: Price List Version in Product | Price tab
The SL_ProductPrice_PriceListVersion callout overwrote the Product Price organization with the price list version organization to ensure the new record had the right organization. However, since the organization field is not shown in that tab, this behavior created a problem because it changed the organization field with the organization of the first price list version in alphabetic order (which was automatically set in the price list version selector). Thus, if the user wanted to select another pricelist from the list, the automatic validation only showed the ones belonging to the selected organization.

The solution is to remove that callout, and to implement the same logic in an EventObserver. Thus the default organization in the Product Price is set to the organization in the Product (as any other child tab in the system), and the ProductPriceObserver ensures the new record belongs to the price list organization when inserting/updating the record.

--HG--
extra : rebase_source : 9f5d73d4660407844b42ea2fc2866f5e4eb9ebcb