Fixed bug 30815: Scheduled Organization for Costing background process
The Costing Background Process should only be scheduled for a legal entity or any of its ancestor.

The fix adds a validation in both the ProcessRequestEventHandler, so each time a record is saved/updated the validation is run, and in the CostingBackground class itself to control wrong process request scheduled before this validation.