Fixed bug 31022: hidden fields in several windows related to acct. dimensions
The display logic of some fields driven by the accounting dimension display logic utility was not working fine when the auxiliary input "DOCBASETYPE" (used by to calculate the acct. dimension display logic) was null.
This auxiliary input has a hardcoded value in some windows, however in other windows it is gotten from the DB using the selected document type. In the latter scenario, when the document type field is blank, the "DOCBASETYPE" auxiliary input was null, so the system was unable to detect the right configuration, thus hiding the fields.

The regression is caused by issue #29767. The business partner and organization fields were configured to use the accounting dimension display logic utility in the following windows:
* Sales Order
* Purchase Order
* Sales Invoice
* Purchase Invoice
* Payment In (not affected as it has a hardcoded "DOCBASETYPE")
* Payment Out (not affected as it has a hardcoded "DOCBASETYPE")

The fix improves the SQL for getting the "DOCBASETYPE" auxiliary input in the affected windows, so in case the document type is null, a generic value (the most commonly used) is returned depending on the window:
* Sales Order: SOO
* Purchase Order: POO
* Sales Invoice: ARI
* Purchase Invoice: API

--HG--
extra : transplant_source : Fo%C3%81%12%D4%B6F%90%F0%FA%03%F5%22%11%07%0B%E9%C4%87