Fixed issue 31056: Create a new record in a form was not working fine.
The problem was that, if you create a new record in form view, and after fulfilling the form, you
click in the create a new record in a form button without previously saving the firstly created record,
it was not working fine. In Goods Movements lines tab, for example, the value of the product chosen in
the first record was shown. That wasn't correct, the product should be empty.

The problem was in the "ob-standard-view-datasource.js" the value map cache was being invalidated.

To fix this, some clases have been changed.
In the "ob-standard-view.js" in the "newDocument" function the parameters the isNewDocument parameter
has been added and it is set to true.
This parameter is passed to the request for being able to take it in the class "ob-standard-view-datasource.js"
and to use it in the logic that invalidates the value map cache.
Now, if this parameter is set to true, the value map cache is not invalidated.