Fixed bug 31039: Cannot create sales order from project phase if the customer default invoice terms is Contract Billing
The v_InvoiceRule variable was defined as CHAR. In case the C_BPartner.InvoiceRule length is higher than 1 (as when defining Contract Billing), the process throws an exception.
The solution is to declare the variable as C_BPartner.InvoiceRule%TYPE, which automatically takes the definition from the C_BPartner table.