Fixed issue 30933: Selectors with tree reference were not working fine.
Some classes have been changed; first of all, in the ob-formitem-tree-filter.js the problem was that there was no logic
for the selector windows and this logic has been added.

There was other problem, in SelectorComponent.java. This class is the one which handles the ftl of the selectors and
there was no logic for the tree references. The getFilterEditorProperties() has been modified for being able to show
the needed structure in the view. The Datasource Id, the treeGridFields and the treeReference were needed.
A new method called getTreeGridFieldsDefinition() has been created to get the correct structure with the correct data in the
tree grid fields.

Now it is possible to use a new selector with a column whose reference is a tree.

--HG--
extra : rebase_source : 05a541293f9dd8dd5ac8f95f6d61e0564aaa9208