Fixed bug 31168: Copy Version process in Process Plan window is not working in Oracle
Ad_Sequence_Doc is a stored procedure that returns the next sequence number using an "out" parameter. The CallStoredProcedure.call() used to call that procedure doesn't currently support parameters of type "out", creating an exception in Oracle only (and not in PostgreSQL because the procedure is declared as a function returning the value).

Instead of calling the procedure using CallStoredProcedure.call(), we use Utility.getDocumentNo() method instead, which is a more standard way to get the next sequence no. from Java, and supports Oracle procedures with "out" parameters.

Besides the class has been modified to properly show an error message in the UI in case an exception is raised, which is something not working before neither.