[alfresco] Use FileItem instead of UploadItem in attachment window.
It is not possible to use UploadItem on a form that redraws. This is the case
when display logics have to be executed. FileItem is used instead.

When using a FileItem it is not possible to do submitForm as the file will not
be included in the request. In this case it is used the internal DynamicForm
used by the FileItem. To include all the parameters in the request they need
to be added as hidden inputs in the DynamicForm that it is finally submited.