Related to issue 30986:Code review improvements.
The problem was when creating new language code has been included in AD_LANGUAGE_CREATE
which is in core, being insertions for a table in client.application module, the code
managing them should be within that module (OBUIAPP_LANGUAGE_CREATE) where parameters
are also handled.

Now it has been resolved the problem by moving the code to the correct function
(OBUIAPP_LANGUAGE_CREATE).