Fixes Issue 31176. Changes old selector with new one in M_InoutLine Table.
Replaces the old selector for Storage Bins with a new OBUISEL_Selector
in M_InoutLine Table.
In Sales flow it shows only Storage Bins that are available to the selected
Organization and that are also On Hand Warehouses.
In the Purchase flow it shows Storage Bins that are available to the selected
Organization.
Also, the selector is pre-filtered using the Warehouse in the Document Header
and it has as a default value a Storage Bin of that Warehouse.