Fixed issue 30783: Reduce the number of clicks to print Documents.
Before, when a document was going to be printed, several clicks were needed.
That was because when you click on the print button, a popup which asks if
you wanted to attach the document or just print it appear, and after it was possible to print.

The number of clicks needed for that task  could be bothering if it was often repeated:.

To fix this two new preferences have been added.
The first one "Attach by default": If this preference is set to "N" then, the popup is not
shown when clicking the print button, and the document is not attached.
If this preference is set to "Y" then, the popup is not shown when clicking the print button
and the document is attached.
If this preference is undefined, then, the popup which asks for attaching the document or just printing
it appears when clicking the print button.

The other preference "Direct Print": If this is set as "Y" then the printing popup will be
opened.
If this is set as "N" or undefined, then the document is downloaded, but the printing popup is not shown.
The "ReportingUtils.java" has been modified, now if the direct print preference is set, the "this.print()" js function
is executed.
The "PrinterReports.java" has been modified for handling the new functionality with the preference which
attaches the documents directly.
The "PrintController.java" has been modified for handling the new functionality with the Direct print
preference.

The index.jsp has been changed because there was a problem with some browsers and the "Direct print" preference
it was not possible to display the printer popup in Internet Explorer nor Firefox.
This was because we were using a hidden frame for putting the pdf there. Chrome was able to take what there was in
that frame, but Firefox and Internet Explorer were not able.
To fix this the style="display:none" of the frame "background_target" has been removed.
Now, it is possible to use the Direct Print preference in the Firefox, Safari, Internet Explorer and Chrome browsers.

--HG--
extra : rebase_source : 7fadfade8dc8e76998642c5b226feb62219f699c