Fixed bug 31198:RO logic doesn't work properly in grid view with combo fields
The problem was that Read Only logic did not work properly in grid view with
combo type fields. If there were two fields on a grid, one having 'List' as
a reference and the second one been a 'TableDir' and the second one had a read
only logic based on the value of the first one, the second one did not work
properly. This happend when 'List' field was next to (leftside) of the
'TableDir' field or if there were intermediate fields between 'List' and
'TableDir' fields that were not editables (RO).

The problem has been fixed overriding a 'refreshCell()' method to use
'allowEditCellRefresh' parameter. This parameter is taking into account in
'refreshCellValue()' method of SC and if there is a visible editor in the
cell with the focus on it, redrawing and update it's value.

This fix take into account when 'refreshCell()' method is called with a value
in 'allowEditCellRefresh' parameter and in this case it will not be called
the overridden method to avoid delete some preexisting  correct fluxes.