Fixed issue 31326: Return values are not properly displayed in Web POS
The problem was that before the orderType wasn't set to  '1' (Return). With
previous fixes, now the orderType is set to 1.

The solution is to not execute the function changeSignToShowReturns() from
order.js. It's necessary to change the order type title from the order in the
POS. Finally, it's necessary to change to change the printAmount function to
show the negative valor of the payment.

Now, the information of the return is shown properly.

--HG--
extra : rebase_source : 95f65d74217669afbf3e92ce2afd325c494c07cf