Fixed issue 31317: The error shown in the log is fixed.
There was an error shown in the log because the servlet was interpreting that it was an email.
That was caused because a condition was not taking into account nothing else but the request.getServletPath as
"print.html".
This has been changed and now "printoptions.html" is taken into account.