Related to issue 30986:Problem adding to a role a process definition access
The problem was when it is tried to translated a "process definition". In
a "Role" window it can be added a "process definition" access and it is
not translated.

The process "generate.entities" has thrown an error because a column that
is call process definition in obuiapp_process_trl has "Link to Parent
Column" unselected.

This check had been forgotten to set properly.Now, It has been resolved
with this fix.