Fixed issue 31344: The email option does not work wih attach by default.
When  the attach by default preference was set, the email option was not working.
To fix this some classes have bennchanged:
OBViewTab.java to have a new parameter "emailButton" or "printButton" in the view.
In the ob-toolbar.js, in the OB.ToolbarUtils.print function now takes the new parameter and puts
it in the popupParams.
Thanks to that in the PrinterReports.java it is possible to catch that parameter from the
vars variable and if it is "emailButton" the logic which was failing at first is avoided.
That logic is only for the direct print and the attach by default and this has no sense in the case
of an email.

--HG--
extra : rebase_source : ee9c87f3584e3496c99d55c4682d5e1c2610b9f4