fixes issue 31293: current combo value is not selected when opening drop down
The problem is caused because after the combo refactor project table and tabledir references make use of a class with different behavior (OBFKComboItem).

Together with this, currently in the FormItem class there is a logic on the _shouldSelectOnFocus method which performs the selection just when tabbing through the form, or on programmatic focus, but not on mouse click. This is the main reason why the text is not being selected when clicking on the drop-down in form view for any reference.

In order to have a common behavior for all the combos in both grid and form view, we are overriding the showPicker() method of OBComboBoxItem to ensure that we select the text after clicking on the drop-down