fixes issue 31377: Combo value not selected after open drop-down with keyboard
Now the behavior of the combos when opening the drop-down with the keyboard is the same as when clicking on the arrow button with the mouse.

To solve this problem the handleKeyPress function is overriden to select the current combo value when ALT + keyboard down shortcut is pressed. Also it is checked that the combo is on a valid state before showing the drop-down.

--HG--
extra : rebase_source : 216a67fa9e88ed6a7aff63153e805423c20e8772