Fixes issue 31061: Can not reactivate a invoice if it's related to a LandedCost
It is not possible to reactivate a purchase invoice if any line has a related landed cost matched.
Now, it will be possible to manually remove this landed cost matched, if matching is not processed, by unselecting it in Match LC Costs pick and edit, and then reactivate purchase invoice.