Fixed issue 30806: The default sales representative is by default the current logged user, even if it not a Sales representative
Changed the place where the POS checks if the logged user is a valid
representative from the Context to the login part.

If the current logged user is not a valid sales representative, we set as null
the sales representative filed in the receipt properties.

--HG--
extra : rebase_source : 48e3c47a9e07a4ea18c782d5e5ed379671b6d7a2