Fixed issue 31360: Cash management is using two gl items distinct
Changes made in the ProcessCashMgmt.

This class is the responsible of create two transactions in the backoffice from
the cash management. The first one is the transaction of the cash management.
The second one is an inverted transaction of the cash management.

The problem was that this two transactions must use the same G/L Item.

Now both transactions use the same G/L Item, depending of the cash management.

--HG--
extra : rebase_source : bf65114ae5ef61e94a6a64d6a49268bd26d2a164