fixes issue 31411: Not possible to install modules in Module Mgmt using Edge
The problem is that on MS Edge browser, the window.opener was lost. Now it is recovered in the same way as it was done for IE9. This avoids the javascript error that was being fired when trying to hide the loading screen in that browser.

Together with this the cache revision number has been updated in the required files.

--HG--
extra : rebase_source : 6c6be8ba07542db5fc610a98f34754946aa1e596