fixes issue 31070: Deleting pop-up is never closed under some circumnstances
The problem was that when removing a record that is being editing, the updateTabTitle function was crashing when trying to access to this.viewGrid.getEditForm().isNew because this.viewGrid.getEditForm() is undefined at that point.

Apart from this, the isEditingGrid is not restored to false when removing a record which is being edited. This caused new errors after removing, because the view is not marked as in non-editing state. For this reason, now the isEditingGrid value is set to false for this case also.

--HG--
extra : rebase_source : 010542a441f58e9c35816f4818f45fef30a2cf9a