fixes issue 30436: Some columns appear empty when adding them in a saved view
The problem happened when adding a field with a saved view which by defaul is not displayed in the grid. In that case the values for that field are not present in the client side and is necessary to perform a request to the datasource.

To solve the problem a new method has been included for OBViewGrid, viewHasFieldsNotInGrid, which checks if every field added by the saved view is already displayed in the grid. If not then we invalidate the cache to force a request to the datasource through the refreshContents() function.

--HG--
extra : rebase_source : 55e62183207498261d1010a846f81608f29fe3d7