fixes issue 28977: NPE if property field can not reach the complete path
The NPE was thrown when defining a property field in the header tab. The FIC is using the getValueFromPath() method to calculate the value of the property field on NEW mode. It is sending as parameter the reference to the parent record for the cases where the value of the property field can be retrieved using the parent information. As in the header tabs we do not have a reference to a parent record, the parameter is null in these cases.

To solve the problem now we first check in getValueFromPath() method if the bob passed as parameter is null. In that case we return null in order to avoid the NPE.